Actually call Requests' json method in the response.
Not sure why this worked before. Perhaps because the Response
object has an __iter__ implementation, or perhaps some Python magic.
Anyway, this used to work fine when writing and testing the api,
but when I started using it in the wild, I would get "bound method
'json'" or "'instancemethod' object" with "is not subscriptable".
Which makes sense.

While reading the code of Requests' Response object, I also
discovered that it has a __bool__ magic method implemented
which calls and returns the ok method, so I removed the .ok calls
from the ifs.